# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Rule Engine Service",
    "description": "Generic configurable rule engine backend",
    "version": "0.1.0"
  },
  "paths": {
...
    "/rules/": {
      "get": {
        "tags": [
          "Rules"
        ],
        "summary": "List Rules",
        "description": "Retrieve a list of Rules.\n\nAccess Control:\n    - Only ADMIN and AUTHOR can view rules\n\nQuery Parameters:\n    entity_version_id: Filter by version (optional but recommended)\n    skip: Pagination offset\n    limit: Maximum results (max 100)\n\nReturns:\n    List[RuleRead]: List of rules",
        "operationId": "list_rules_rules__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "entity_version_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Entity Version Id"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Number of records to skip",
              "default": 0,
              "title": "Skip"
            },
            "description": "Number of records to skip"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 0,
              "description": "Maximum number of records to return",
              "default": 100,
              "title": "Limit"
            },
            "description": "Maximum number of records to return"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RuleRead"
                  },
                  "title": "Response List Rules Rules  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Rules"
        ],
        "summary": "Create Rule",
        "description": "Creates a new Rule in a DRAFT version.\n\nRestrictions:\n    - The version must be DRAFT\n    - Target Field must belong to the specified Version\n    - Target Value (if specified) must belong to the Target Field\n\nAccess Control:\n    - Only ADMIN and AUTHOR can create rules\n\nReturns:\n    RuleRead: The created rule",
        "operationId": "create_rule_rules__post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuleCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/rules/{rule_id}": {
      "get": {
        "tags": [
          "Rules"
        ],
        "summary": "Read Rule",
        "description": "Retrieve a single Rule.\n\nAccess Control:\n    - Only ADMIN and AUTHOR can view rule details\n\nReturns:\n    RuleRead: The requested rule",
        "operationId": "read_rule_rules__rule_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "rule_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "description": "Rule ID",
              "title": "Rule Id"
            },
            "description": "Rule ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Rules"
        ],
        "summary": "Update Rule",
        "description": "Updates an existing Rule.\n\nRestrictions:\n    - The version must be DRAFT\n    - Cannot change entity_version_id (rules belong strictly to their creation version)\n    - New target Field must belong to the same Version\n    - New target Value must belong to the new target Field\n\nAccess Control:\n    - Only ADMIN and AUTHOR can update rules\n\nReturns:\n    RuleRead: The updated rule",
        "operationId": "update_rule_rules__rule_id__patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "rule_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "description": "Rule ID",
              "title": "Rule Id"
            },
            "description": "Rule ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuleUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Rules"
        ],
        "summary": "Delete Rule",
        "description": "Delete a Rule.\n\nRestrictions:\n    - The version must be DRAFT\n\nAccess Control:\n    - Only ADMIN and AUTHOR can delete rules\n\nReturns:\n    204 No Content on success",
        "operationId": "delete_rule_rules__rule_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "rule_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "description": "Rule ID",
              "title": "Rule Id"
            },
            "description": "Rule ID"
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/bom-items/": {
      "get": {
        "tags": [
          "BOM Items"
        ],
        "summary": "List Bom Items",
        "description": "Retrieve BOM items for a specific version.\n\nAccess Control:\n    - Only ADMIN and AUTHOR can view BOM items",
        "operationId": "list_bom_items_bom_items__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "entity_version_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Entity Version Id"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Number of records to skip",
              "default": 0,
              "title": "Skip"
            },
            "description": "Number of records to skip"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 0,
              "description": "Maximum number of records to return",
              "default": 100,
              "title": "Limit"
            },
            "description": "Maximum number of records to return"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BOMItemRead"
                  },
                  "title": "Response List Bom Items Bom Items  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "BOM Items"
        ],
        "summary": "Create Bom Item",
        "description": "Create a BOM item attached to a specific entity version.\n\nRestrictions:\n    - The version must be DRAFT\n    - Pricing validation by bom_type\n    - quantity must be > 0\n    - quantity_from_field_id must reference a NUMBER field in the same version\n    - parent_bom_item_id must reference an item in the same version\n\nAccess Control:\n    - Only ADMIN and AUTHOR can create BOM items",
        "operationId": "create_bom_item_bom_items__post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BOMItemCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BOMItemRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/bom-items/{bom_item_id}": {
      "get": {
        "tags": [
          "BOM Items"
        ],
        "summary": "Read Bom Item",
        "description": "Retrieve a single BOM item.\n\nAccess Control:\n    - Only ADMIN and AUTHOR can view BOM item details",
        "operationId": "read_bom_item_bom_items__bom_item_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "bom_item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "description": "BOM Item ID",
              "title": "Bom Item Id"
            },
            "description": "BOM Item ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BOMItemRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "BOM Items"
        ],
        "summary": "Update Bom Item",
        "description": "Update a BOM item.\n\nRestrictions:\n    - The version must be DRAFT\n    - Pricing validation applies on type change\n\nAccess Control:\n    - Only ADMIN and AUTHOR can update BOM items",
        "operationId": "update_bom_item_bom_items__bom_item_id__patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "bom_item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "description": "BOM Item ID",
              "title": "Bom Item Id"
            },
            "description": "BOM Item ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BOMItemUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BOMItemRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "BOM Items"
        ],
        "summary": "Delete Bom Item",
        "description": "Delete a BOM item and its children (cascade).\n\nRestrictions:\n    - The version must be DRAFT\n\nAccess Control:\n    - Only ADMIN and AUTHOR can delete BOM items",
        "operationId": "delete_bom_item_bom_items__bom_item_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "bom_item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "description": "BOM Item ID",
              "title": "Bom Item Id"
            },
            "description": "BOM Item ID"
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/bom-item-rules/": {
      "get": {
        "tags": [
          "BOM Item Rules"
        ],
        "summary": "List Bom Item Rules",
        "description": "Retrieve BOM item rules filtered by bom_item_id or entity_version_id.\n\nAt least one filter parameter must be provided.\n\nAccess Control:\n    - Only ADMIN and AUTHOR can view BOM item rules",
        "operationId": "list_bom_item_rules_bom_item_rules__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "bom_item_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Bom Item Id"
            }
          },
          {
            "name": "entity_version_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Entity Version Id"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Number of records to skip",
              "default": 0,
              "title": "Skip"
            },
            "description": "Number of records to skip"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 0,
              "description": "Maximum number of records to return",
              "default": 100,
              "title": "Limit"
            },
            "description": "Maximum number of records to return"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BOMItemRuleRead"
                  },
                  "title": "Response List Bom Item Rules Bom Item Rules  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "BOM Item Rules"
        ],
        "summary": "Create Bom Item Rule",
        "description": "Create a BOM item rule attached to a specific BOM item.\n\nRestrictions:\n    - The version must be DRAFT\n    - bom_item_id must belong to the specified entity_version_id\n    - All field_id values in conditions.criteria must belong to the same entity_version_id\n\nAccess Control:\n    - Only ADMIN and AUTHOR can create BOM item rules",
        "operationId": "create_bom_item_rule_bom_item_rules__post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BOMItemRuleCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BOMItemRuleRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "RuleConditions": {
        "properties": {
          "criteria": {
            "items": {
              "$ref": "#/components/schemas/RuleCriterion"
            },
            "type": "array",
            "title": "Criteria"
          }
        },
        "type": "object",
        "required": [
          "criteria"
        ],
        "title": "RuleConditions"
      },
      "RuleCreate": {
        "properties": {
          "conditions": {
            "$ref": "#/components/schemas/RuleConditions"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "rule_type": {
            "$ref": "#/components/schemas/RuleType",
            "default": "availability"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "set_value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Set Value"
          },
          "entity_version_id": {
            "type": "integer",
            "title": "Entity Version Id"
          },
          "target_field_id": {
            "type": "integer",
            "title": "Target Field Id"
          },
          "target_value_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Value Id"
          }
        },
        "type": "object",
        "required": [
          "conditions",
          "entity_version_id",
          "target_field_id"
        ],
        "title": "RuleCreate",
        "description": "Payload to create a new version."
      },
...
      "RuleRead": {
        "properties": {
          "conditions": {
            "$ref": "#/components/schemas/RuleConditions"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "rule_type": {
            "$ref": "#/components/schemas/RuleType",
            "default": "availability"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "set_value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Set Value"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "entity_version_id": {
            "type": "integer",
            "title": "Entity Version Id"
          },
          "target_field_id": {
            "type": "integer",
            "title": "Target Field Id"
          },
          "target_value_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Value Id"
          }
        },
        "type": "object",
        "required": [
          "conditions",
          "id",
          "entity_version_id",
          "target_field_id"
        ],
        "title": "RuleRead",
        "description": "Output schema for API responses."
      },
...
      "RuleUpdate": {
        "properties": {
          "conditions": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RuleConditions"
              },
              {
                "type": "null"
              }
            ]
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "rule_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RuleType"
              },
              {
                "type": "null"
              }
            ]
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "set_value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Set Value"
          },
          "target_field_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Field Id"
          },
          "target_value_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Value Id"
          }
        },
        "type": "object",
        "title": "RuleUpdate",
        "description": "Schema for updating existing rule (PATCH). All fields optional.\n\nNote: entity_version_id is not included because moving rules between versions is forbidden.\nFull consistency validation (rule_type vs target_value_id) is performed at the service layer\nwhere the existing rule data is available."
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "auth/token"
          }
        }
      }
    }
  }
}